+++ NexaRun.Cli/Commands/StopCommand.cs
@@ -9,17 +9,13 @@ public static class StopCommand
 {
     public static Command Build(IpcClient client)
     {
-        var nameArg = new Argument<string>("name")
-        {
-            Description = "Name of the process to stop"
-        };
-
-        var cmd = new Command("stop", "Stop a running process") { nameArg };
+        var targetArg = CliCommands.TargetArgument("stop");
+        var cmd = new Command("stop", "Stop a running process by id or name") { targetArg };
 
         cmd.SetAction(async parseResult =>
         {
-            var name = parseResult.GetValue(nameArg)!;
-            var response = await client.Send(new IpcRequest { Command = "stop", ProcessName = name });
+            var target = parseResult.GetValue(targetArg)!;
+            var response = await client.Send(CliCommands.TargetRequest("stop", target));
             return CliOutput.Exit(response);
         });